+++ package.json
@@ -56,6 +56,7 @@
     "@tailwindcss/forms": "^0.5.6",
     "@testing-library/jest-dom": "^6.1.3",
     "@testing-library/react": "^14.0.0",
+    "@types/gtag.js": "^0.0.14",
     "@types/jest": "^29.5.5",
     "@types/qs": "^6.9.8",
     "@types/react": "^18.2.22",
