+++ src/components/layout/GoogleAnalytics.tsx
@@ -1,0 +1,45 @@
+import { usePathname, useSearchParams } from 'next/navigation';
+import { useEffect } from 'react';
+
+import { pageview } from '@/lib/helper';
+
+export default function GoogleAnalytics({
+  GA_MEASUREMENT_ID,
+}: {
+  GA_MEASUREMENT_ID: string;
+}) {
+  const pathname = usePathname();
+  const searchParams = useSearchParams();
+
+  useEffect(() => {
+    const url = pathname + searchParams.toString();
+
+    pageview(GA_MEASUREMENT_ID, url);
+  }, [pathname, searchParams, GA_MEASUREMENT_ID]);
+
+  return (
+    <>
+      {/* eslint-disable-next-line @next/next/no-sync-scripts */}
+      <script
+        src={`https://www.googletagmanager.com/gtag/js?id=${GA_MEASUREMENT_ID}`}
+        type='text/plain'
+        data-cookieconsent='statistics'
+      />
+      <script
+        id='google-analytics'
+        type='text/plain'
+        data-cookieconsent='statistics'
+      >
+        {`
+          window.dataLayer = window.dataLayer || [];
+          function gtag(){dataLayer.push(arguments);}
+          gtag('js', new Date());
+
+          gtag('config', '${GA_MEASUREMENT_ID}', {
+            page_path: window.location.pathname,
+          });
+        `}
+      </script>
+    </>
+  );
+}
