+++ src/lib/helper.ts
@@ -68,6 +68,12 @@ export function formatDate(dateString: Date) {
   return date.toLocaleDateString('de-DE', options);
 }
 
+export const pageview = (GA_MEASUREMENT_ID: string, url: string) => {
+  window.gtag('config', GA_MEASUREMENT_ID, {
+    page_path: url,
+  });
+};
+
 // ADDS DELAY TO SIMULATE SLOW API REMOVE FOR PRODUCTION
 export const delay = (time: number) =>
   new Promise((resolve) => setTimeout(() => resolve(1), time));
